test: add test for SettingsCard
